[cleanup] moves App.js to containers
